requirements.txt
Django==2.2.24
django-shopping-cart==0.1
pytz==2021.1
sqlparse==0.4.1
